requirements.txt
pytest>=5.0.1
PyYAML>=5.1.2
requests>=2.22.0